show-environment: Add posix output mode.
Add a -p option to show-environment to output the environment variables
in a form that can be eval'ed by a posix-compatible shell such that the
shell environment will reflect the session/global environment.

Set variables are output as 'VAR="VAL"; export VAR;' (with appropriate
escaping for special characters inside the double quotes), and unset
variables are output as 'unset VAL;'.

This makes is simple to propagate the session environment to an existing
shell in a tmux window by running 'eval $(tmux show-environment -p)'.