copy-mode: Add copy-mode option to exit on scroll.
Add a -e option that will make copy-mode exit when scrolling to
the bottom of the history, if there is no selection in progress.
This re-enables the ability for mouse scrolling to automatically exit
when scrolling to the bottom (remove in mouse handling refactor). It
also makes quick scroll buffer preview easy allowing easy use of
NPage and PPage to enter and exit copy mode as necessary.